thesis_backend/requirements.txt
amqp==5.1.1
autopep8==2.0.2
billiard==4.1.0
celery==5.3.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4
contourpy==1.0.7
cycler==0.11.0
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.2.3
Flask-CeleryExt==0.5.0
Flask-JWT-Extended==4.5.3
flask-mongoengine==1.0.0
Flask-WTF==1.1.1
fonttools==4.39.3
globals==0.3.36
gunicorn==21.2.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
kombu==5.3.2
MarkupSafe==2.1.2
matplotlib==3.7.1
mongoengine==0.27.0
mysql-connector-python==8.0.33
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.5.0
Pint==0.20.1
pluggy==1.0.0
prompt-toolkit==3.0.39
psutil==5.9.5
pycodestyle==2.10.0
PyJWT==2.8.0
PyMeasure==0.11.1
pymongo==4.3.3
pyparsing==3.0.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyqtgraph==0.13.2
pyserial==3.5
pytest==7.3.1
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-helper==0.3.65
pytz==2023.2
PyVISA==1.13.0
six==1.16.0
typing_extensions==4.5.0
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
WTForms==3.0.1
